backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.1
httpx>=0.28.0
apscheduler>=3.10.4
